refactor(dashboard): extract catalog text resolvers into a shared i18n module
resolveTitle / resolveSubtitle / seriesEntryLabel move out of KpiTile.jsx
into products/dashboard/src/i18n/textResolver.js — the descriptor-text
companion of dimensionLabel.js — so every component resolves catalog text
through the same seam. KpiTile, DashboardTable (threshold tag labels) and
TableSortHeader (column labels) now consume it; no behavior change.
