+++ digit-ui-esbuild/products/dashboard/src/components/KpiTile.jsx
@@ -20,7 +20,8 @@ import {
 } from '../config/complaintsAtRiskPresentation';
 import useDashboardT from '../i18n/useDashboardT';
 import { dimensionLabel } from '../i18n/dimensionLabel';
-import { translate as t, exists } from '../i18n/localeRuntime';
+import { translate as t } from '../i18n/localeRuntime';
+import { resolveTitle, resolveSubtitle, seriesEntryLabel } from '../i18n/textResolver';
 
 /**
  * Generic viz-kind-driven tile renderer (the dashboard RENDERING ENGINE).
@@ -1074,54 +1075,6 @@ function epochOrIsoToDateKey(value) {
 
 function normalizeSeg(value) {
   return String(value ?? '').toUpperCase();
-}
-
-/**
- * Title resolution for the inverted catalog. `titleKey` (a raw
- * RAINMAKER-PGR.DASHBOARD_KPI_* localization code) wins when its message is
- * seeded; otherwise the human `viz.title` from the MDMS def is the source of
- * truth, and the prettified key remains the last-resort fallback — never
- * rendered verbatim.
- */
-function resolveTitle(def) {
-  const titleKey = def?.viz?.titleKey || def?.titleKey;
-  if (titleKey && exists(titleKey)) return t(titleKey);
-  return (
-    def?.viz?.title ||
-    def?.title ||
-    def?.name ||
-    prettifyTitleKey(titleKey) ||
-    ''
-  );
-}
-
-/**
- * Subtitle: `viz.subtitleKey` (RAINMAKER-PGR.DASHBOARD_KPI_<ID>_SUBTITLE) wins
- * when seeded, else the legacy viz.subtitle / description / contextLabel chain.
- */
-function resolveSubtitle(viz) {
-  const subtitle = viz.subtitleKey && exists(viz.subtitleKey) ? t(viz.subtitleKey) : viz.subtitle;
-  return subtitle || viz.description || viz.contextLabel || '';
-}
-
-/**
- * stackSeries / seriesDefs / channelMap / measure-column entries: a `labelKey`
- * (DASHBOARD_WF_STAGE_* / DASHBOARD_CHANNEL_* / DASHBOARD_SLA_* / DASHBOARD_COL_*)
- * wins when seeded, else the descriptor's literal label.
- */
-function seriesEntryLabel(entry, fallback) {
-  return entry?.labelKey && exists(entry.labelKey) ? t(entry.labelKey) : fallback;
-}
-
-function prettifyTitleKey(key) {
-  if (!key) return '';
-  const tail = String(key).split('.').pop().replace(/^DASHBOARD_KPI_/, '');
-  if (!tail) return '';
-  return tail
-    .toLowerCase()
-    .replace(/_/g, ' ')
-    .replace(/\b\w/g, (c) => c.toUpperCase())
-    .trim();
 }
 
 function errorLabel(code) {
